Implemented Lukas' suggestions
