+++ .env.template
@@ -4,6 +4,4 @@ POSTGRES_DB=qasm
 POSTGRES_PORT=5432
 POSTGRES_HOST=postgres
 SQLALCHEMY_DRIVER=postgresql+psycopg
-API_PROTOCOL=http
-API_DOMAIN=localhost
-API_PORT=8000
+API_BASE_URL=http://localhost:8000/
